// src/app/service/model/form1Dto.ts
/**
 * Swagger First Rest Api æ–‡æ¡£
 * user æ¨¡å—apiæ–‡æ¡£
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Form1SubDto } from './form1SubDto';


export interface Form1Dto { 
    faddr?: string;
    fhouseType?: string;
    fidNo?: string;
    fname?: string;
    fphone?: string;
    freturnDate?: Date;
    freturnDlPath?: string;
    fsex?: string;
    id?: number;
    returnPathList?: Array<Form1SubDto>;
}