// src/app/service/model/form1SubDto.ts
/**
 * Swagger First Rest Api æ–‡æ¡£
 * user æ¨¡å—apiæ–‡æ¡£
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Form1SubDto { 
    from?: string;
    target?: string;
    trafficDate?: Date;
    trafficNo?: string;
    trafficType?: string;
}